Struct aws_sdk_kms::client::fluent_builders::Encrypt

pub struct Encrypt<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> Encrypt<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<EncryptOutput, SdkError<EncryptError>> where
R::Policy: SmithyRetryPolicy<EncryptInputOperationOutputAlias, EncryptOutput, EncryptError, EncryptInputOperationRetryAlias>,

pub fn key_id(self, inp: impl Into<String>) -> Self

Identifies the KMS key to use in the encryption operation.

To specify a KMS key, use its key ID, key ARN, alias name, or alias ARN. When using an alias name, prefix it with "alias/". To specify a KMS key in a different Amazon Web Services account, you must use the key ARN or alias ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

To get the key ID and key ARN for a KMS key, use ListKeys or DescribeKey. To get the alias name and alias ARN, use ListAliases.

pub fn set_key_id(self, input: Option<String>) -> Self

pub fn plaintext(self, inp: Blob) -> Self

Data to be encrypted.

pub fn set_plaintext(self, input: Option<Blob>) -> Self

pub fn encryption_context(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

Adds a key-value pair to EncryptionContext.

To override the contents of this collection use set_encryption_context.

Specifies the encryption context that will be used to encrypt the data. An encryption context is valid only for cryptographic operations with a symmetric KMS key. The standard asymmetric encryption algorithms that KMS uses do not support an encryption context.

An encryption context is a collection of non-secret key-value pairs that represents additional authenticated data. When you use an encryption context to encrypt data, you must specify the same (an exact case-sensitive match) encryption context to decrypt the data. An encryption context is optional when encrypting with a symmetric KMS key, but it is highly recommended.

For more information, see Encryption Context in the Key Management Service Developer Guide.

pub fn set_encryption_context(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn grant_tokens(self, inp: impl Into<String>) -> Self

Appends an item to GrantTokens.

To override the contents of this collection use set_grant_tokens.

A list of grant tokens.

Use a grant token when your permission to call this operation comes from a new grant that has not yet achieved eventual consistency. For more information, see Grant token and Using a grant token in the Key Management Service Developer Guide.

pub fn set_grant_tokens(self, input: Option<Vec<String>>) -> Self

pub fn encryption_algorithm(self, inp: EncryptionAlgorithmSpec) -> Self

Specifies the encryption algorithm that KMS will use to encrypt the plaintext message. The algorithm must be compatible with the KMS key that you specify.

This parameter is required only for asymmetric KMS keys. The default value, SYMMETRIC_DEFAULT, is the algorithm used for symmetric KMS keys. If you are using an asymmetric KMS key, we recommend RSAES_OAEP_SHA_256.

pub fn set_encryption_algorithm(
self,
input: Option<EncryptionAlgorithmSpec>
) -> Self