Struct aws_sdk_kms::client::fluent_builders::PutKeyPolicy

pub struct PutKeyPolicy { /* fields omitted */ }

Implementations

impl PutKeyPolicy

pub async fn send(
self
) -> Result<PutKeyPolicyOutput, SdkError<PutKeyPolicyError>>

pub fn key_id(self, inp: impl Into<String>) -> Self

A unique identifier for the customer master key (CMK).

Specify the key ID or the Amazon Resource Name (ARN) of the CMK.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

To get the key ID and key ARN for a CMK, use ListKeys or DescribeKey.

pub fn set_key_id(self, inp: Option<String>) -> Self

pub fn policy_name(self, inp: impl Into<String>) -> Self

The name of the key policy. The only valid value is default.

pub fn set_policy_name(self, inp: Option<String>) -> Self

pub fn policy(self, inp: impl Into<String>) -> Self

The key policy to attach to the CMK.

The key policy must meet the following criteria:

If you don't set BypassPolicyLockoutSafetyCheck to true, the key policy must allow the principal that is making the PutKeyPolicy request to make a subsequent PutKeyPolicy request on the CMK. This reduces the risk that the CMK becomes unmanageable. For more information, refer to the scenario in the Default Key Policy section of the AWS Key Management Service Developer Guide.

Each statement in the key policy must contain one or more principals. The principals in the key policy must exist and be visible to AWS KMS. When you create a new AWS principal (for example, an IAM user or role), you might need to enforce a delay before including the new principal in a key policy because the new principal might not be immediately visible to AWS KMS. For more information, see Changes that I make are not always immediately visible in the AWS Identity and Access Management User Guide.

The key policy cannot exceed 32 kilobytes (32768 bytes). For more information, see Resource Quotas in the AWS Key Management Service Developer Guide.

pub fn set_policy(self, inp: Option<String>) -> Self

pub fn bypass_policy_lockout_safety_check(self, inp: bool) -> Self

A flag to indicate whether to bypass the key policy lockout safety check.

Setting this value to true increases the risk that the CMK becomes unmanageable. Do not set this value to true indiscriminately.

For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide.

Use this parameter only when you intend to prevent the principal that is making the request from making a subsequent PutKeyPolicy request on the CMK.

The default value is false.

pub fn set_bypass_policy_lockout_safety_check(self, inp: bool) -> Self