Struct aws_sdk_kms::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for TrentService.

This client allows ergonomic access to a TrentService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn cancel_key_deletion(&self) -> CancelKeyDeletion<C, M, R>

pub fn connect_custom_key_store(&self) -> ConnectCustomKeyStore<C, M, R>

pub fn create_alias(&self) -> CreateAlias<C, M, R>

pub fn create_custom_key_store(&self) -> CreateCustomKeyStore<C, M, R>

pub fn create_grant(&self) -> CreateGrant<C, M, R>

pub fn create_key(&self) -> CreateKey<C, M, R>

pub fn decrypt(&self) -> Decrypt<C, M, R>

pub fn delete_alias(&self) -> DeleteAlias<C, M, R>

pub fn delete_custom_key_store(&self) -> DeleteCustomKeyStore<C, M, R>

pub fn delete_imported_key_material(&self) -> DeleteImportedKeyMaterial<C, M, R>

pub fn describe_custom_key_stores(&self) -> DescribeCustomKeyStores<C, M, R>

pub fn describe_key(&self) -> DescribeKey<C, M, R>

pub fn disable_key(&self) -> DisableKey<C, M, R>

pub fn disable_key_rotation(&self) -> DisableKeyRotation<C, M, R>

pub fn disconnect_custom_key_store(&self) -> DisconnectCustomKeyStore<C, M, R>

pub fn enable_key(&self) -> EnableKey<C, M, R>

pub fn enable_key_rotation(&self) -> EnableKeyRotation<C, M, R>

pub fn encrypt(&self) -> Encrypt<C, M, R>

pub fn generate_data_key(&self) -> GenerateDataKey<C, M, R>

pub fn generate_data_key_pair(&self) -> GenerateDataKeyPair<C, M, R>

pub fn generate_data_key_pair_without_plaintext(
&self
) -> GenerateDataKeyPairWithoutPlaintext<C, M, R>

pub fn generate_data_key_without_plaintext(
&self
) -> GenerateDataKeyWithoutPlaintext<C, M, R>

pub fn generate_random(&self) -> GenerateRandom<C, M, R>

pub fn get_key_policy(&self) -> GetKeyPolicy<C, M, R>

pub fn get_key_rotation_status(&self) -> GetKeyRotationStatus<C, M, R>

pub fn get_parameters_for_import(&self) -> GetParametersForImport<C, M, R>

pub fn get_public_key(&self) -> GetPublicKey<C, M, R>

pub fn import_key_material(&self) -> ImportKeyMaterial<C, M, R>

pub fn list_aliases(&self) -> ListAliases<C, M, R>

pub fn list_grants(&self) -> ListGrants<C, M, R>

pub fn list_key_policies(&self) -> ListKeyPolicies<C, M, R>

pub fn list_keys(&self) -> ListKeys<C, M, R>

pub fn list_resource_tags(&self) -> ListResourceTags<C, M, R>

pub fn list_retirable_grants(&self) -> ListRetirableGrants<C, M, R>

pub fn put_key_policy(&self) -> PutKeyPolicy<C, M, R>

pub fn re_encrypt(&self) -> ReEncrypt<C, M, R>

pub fn replicate_key(&self) -> ReplicateKey<C, M, R>

pub fn retire_grant(&self) -> RetireGrant<C, M, R>

pub fn revoke_grant(&self) -> RevokeGrant<C, M, R>

pub fn schedule_key_deletion(&self) -> ScheduleKeyDeletion<C, M, R>

pub fn sign(&self) -> Sign<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_alias(&self) -> UpdateAlias<C, M, R>

pub fn update_custom_key_store(&self) -> UpdateCustomKeyStore<C, M, R>

pub fn update_key_description(&self) -> UpdateKeyDescription<C, M, R>

pub fn update_primary_region(&self) -> UpdatePrimaryRegion<C, M, R>

pub fn verify(&self) -> Verify<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self