Module aws_sdk_kms::error

Modules

already_exists_exception

See AlreadyExistsException

cloud_hsm_cluster_in_use_exception

See CloudHsmClusterInUseException

cloud_hsm_cluster_invalid_configuration_exception

See CloudHsmClusterInvalidConfigurationException

cloud_hsm_cluster_not_active_exception

See CloudHsmClusterNotActiveException

cloud_hsm_cluster_not_found_exception

See CloudHsmClusterNotFoundException

cloud_hsm_cluster_not_related_exception

See CloudHsmClusterNotRelatedException

custom_key_store_has_cm_ks_exception

See CustomKeyStoreHasCmKsException

custom_key_store_invalid_state_exception

See CustomKeyStoreInvalidStateException

custom_key_store_name_in_use_exception

See CustomKeyStoreNameInUseException

custom_key_store_not_found_exception

See CustomKeyStoreNotFoundException

dependency_timeout_exception

See DependencyTimeoutException

disabled_exception

See DisabledException

expired_import_token_exception

See ExpiredImportTokenException

incorrect_key_exception

See IncorrectKeyException

incorrect_key_material_exception

See IncorrectKeyMaterialException

incorrect_trust_anchor_exception

See IncorrectTrustAnchorException

invalid_alias_name_exception

See InvalidAliasNameException

invalid_arn_exception

See InvalidArnException

invalid_ciphertext_exception

See InvalidCiphertextException

invalid_grant_id_exception

See InvalidGrantIdException

invalid_grant_token_exception

See InvalidGrantTokenException

invalid_import_token_exception

See InvalidImportTokenException

invalid_key_usage_exception

See InvalidKeyUsageException

invalid_marker_exception

See InvalidMarkerException

key_unavailable_exception

See KeyUnavailableException

kms_internal_exception

See KmsInternalException

kms_invalid_signature_exception

See KmsInvalidSignatureException

kms_invalid_state_exception

See KmsInvalidStateException

limit_exceeded_exception

See LimitExceededException

malformed_policy_document_exception

See MalformedPolicyDocumentException

not_found_exception

See NotFoundException

tag_exception

See TagException

unsupported_operation_exception

See UnsupportedOperationException

Structs

AlreadyExistsException

The request was rejected because it attempted to create a resource that already

CancelKeyDeletionError

CloudHsmClusterInUseException

The request was rejected because the specified AWS CloudHSM cluster is already associated with a

CloudHsmClusterInvalidConfigurationException

The request was rejected because the associated AWS CloudHSM cluster did not meet the

CloudHsmClusterNotActiveException

The request was rejected because the AWS CloudHSM cluster that is associated with the custom key

CloudHsmClusterNotFoundException

The request was rejected because AWS KMS cannot find the AWS CloudHSM cluster with the specified

CloudHsmClusterNotRelatedException

The request was rejected because the specified AWS CloudHSM cluster has a different cluster

ConnectCustomKeyStoreError

CreateAliasError

CreateCustomKeyStoreError

CreateGrantError

CreateKeyError

CustomKeyStoreHasCmKsException

The request was rejected because the custom key store contains AWS KMS customer master keys

CustomKeyStoreInvalidStateException

The request was rejected because of the ConnectionState of the custom key

CustomKeyStoreNameInUseException

The request was rejected because the specified custom key store name is already assigned

CustomKeyStoreNotFoundException

The request was rejected because AWS KMS cannot find a custom key store with the specified

DecryptError

DeleteAliasError

DeleteCustomKeyStoreError

DeleteImportedKeyMaterialError

DependencyTimeoutException

The system timed out while trying to fulfill the request. The request can be

DescribeCustomKeyStoresError

DescribeKeyError

DisableKeyError

DisableKeyRotationError

DisabledException

The request was rejected because the specified CMK is not enabled.

DisconnectCustomKeyStoreError

EnableKeyError

EnableKeyRotationError

EncryptError

ExpiredImportTokenException

The request was rejected because the specified import token is expired. Use GetParametersForImport to get a new import token and public key, use the new

GenerateDataKeyError

GenerateDataKeyPairError

GenerateDataKeyPairWithoutPlaintextError

GenerateDataKeyWithoutPlaintextError

GenerateRandomError

GetKeyPolicyError

GetKeyRotationStatusError

GetParametersForImportError

GetPublicKeyError

ImportKeyMaterialError

IncorrectKeyException

The request was rejected because the specified CMK cannot decrypt the data. The

IncorrectKeyMaterialException

The request was rejected because the key material in the request is, expired, invalid, or

IncorrectTrustAnchorException

The request was rejected because the trust anchor certificate in the request is not the

InvalidAliasNameException

The request was rejected because the specified alias name is not valid.

InvalidArnException

The request was rejected because a specified ARN, or an ARN in a key policy, is not

InvalidCiphertextException

From the Decrypt or ReEncrypt operation, the request

InvalidGrantIdException

The request was rejected because the specified GrantId is not valid.

InvalidGrantTokenException

The request was rejected because the specified grant token is not valid.

InvalidImportTokenException

The request was rejected because the provided import token is invalid or is associated

InvalidKeyUsageException

The request was rejected for one of the following reasons:

InvalidMarkerException

The request was rejected because the marker that specifies where pagination should next

KeyUnavailableException

The request was rejected because the specified CMK was not available. You can retry the

KmsInternalException

The request was rejected because an internal exception occurred. The request can be

KmsInvalidSignatureException

The request was rejected because the signature verification failed. Signature verification

KmsInvalidStateException

The request was rejected because the state of the specified resource is not valid for this

LimitExceededException

The request was rejected because a quota was exceeded. For more information, see Quotas in the

ListAliasesError

ListGrantsError

ListKeyPoliciesError

ListKeysError

ListResourceTagsError

ListRetirableGrantsError

MalformedPolicyDocumentException

The request was rejected because the specified policy is not syntactically or semantically

NotFoundException

The request was rejected because the specified entity or resource could not be

PutKeyPolicyError

ReEncryptError

ReplicateKeyError

RetireGrantError

RevokeGrantError

ScheduleKeyDeletionError

SignError

TagException

The request was rejected because one or more tags are not valid.

TagResourceError

UnsupportedOperationException

The request was rejected because a specified parameter is not supported or a specified

UntagResourceError

UpdateAliasError

UpdateCustomKeyStoreError

UpdateKeyDescriptionError

UpdatePrimaryRegionError

VerifyError

Enums

CancelKeyDeletionErrorKind

ConnectCustomKeyStoreErrorKind

CreateAliasErrorKind

CreateCustomKeyStoreErrorKind

CreateGrantErrorKind

CreateKeyErrorKind

DecryptErrorKind

DeleteAliasErrorKind

DeleteCustomKeyStoreErrorKind

DeleteImportedKeyMaterialErrorKind

DescribeCustomKeyStoresErrorKind

DescribeKeyErrorKind

DisableKeyErrorKind

DisableKeyRotationErrorKind

DisconnectCustomKeyStoreErrorKind

EnableKeyErrorKind

EnableKeyRotationErrorKind

EncryptErrorKind

GenerateDataKeyErrorKind

GenerateDataKeyPairErrorKind

GenerateDataKeyPairWithoutPlaintextErrorKind

GenerateDataKeyWithoutPlaintextErrorKind

GenerateRandomErrorKind

GetKeyPolicyErrorKind

GetKeyRotationStatusErrorKind

GetParametersForImportErrorKind

GetPublicKeyErrorKind

ImportKeyMaterialErrorKind

ListAliasesErrorKind

ListGrantsErrorKind

ListKeyPoliciesErrorKind

ListKeysErrorKind

ListResourceTagsErrorKind

ListRetirableGrantsErrorKind

PutKeyPolicyErrorKind

ReEncryptErrorKind

ReplicateKeyErrorKind

RetireGrantErrorKind

RevokeGrantErrorKind

ScheduleKeyDeletionErrorKind

SignErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAliasErrorKind

UpdateCustomKeyStoreErrorKind

UpdateKeyDescriptionErrorKind

UpdatePrimaryRegionErrorKind

VerifyErrorKind