Struct aws_sdk_kms::input::create_grant_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateGrantInput

Implementations

impl Builder

pub fn key_id(self, input: impl Into<String>) -> Self

Identifies the customer master key (CMK) for the grant. The grant gives principals permission to use this CMK.

Specify the key ID or key ARN of the CMK. To specify a CMK in a different AWS account, you must use the key ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

To get the key ID and key ARN for a CMK, use ListKeys or DescribeKey.

pub fn set_key_id(self, input: Option<String>) -> Self

pub fn grantee_principal(self, input: impl Into<String>) -> Self

The identity that gets the permissions specified in the grant.

To specify the principal, use the Amazon Resource Name (ARN) of an AWS principal. Valid AWS principals include AWS accounts (root), IAM users, IAM roles, federated users, and assumed role users. For examples of the ARN syntax to use for specifying a principal, see AWS Identity and Access Management (IAM) in the Example ARNs section of the AWS General Reference.

pub fn set_grantee_principal(self, input: Option<String>) -> Self

pub fn retiring_principal(self, input: impl Into<String>) -> Self

The principal that is given permission to retire the grant by using RetireGrant operation.

To specify the principal, use the Amazon Resource Name (ARN) of an AWS principal. Valid AWS principals include AWS accounts (root), IAM users, federated users, and assumed role users. For examples of the ARN syntax to use for specifying a principal, see AWS Identity and Access Management (IAM) in the Example ARNs section of the AWS General Reference.

pub fn set_retiring_principal(self, input: Option<String>) -> Self

pub fn operations(self, input: impl Into<GrantOperation>) -> Self

pub fn set_operations(self, input: Option<Vec<GrantOperation>>) -> Self

pub fn constraints(self, input: GrantConstraints) -> Self

Specifies a grant constraint.

AWS KMS supports the EncryptionContextEquals and EncryptionContextSubset grant constraints. Each constraint value can include up to 8 encryption context pairs. The encryption context value in each constraint cannot exceed 384 characters.

These grant constraints allow a cryptographic operation only when the encryption context in the request matches (EncryptionContextEquals) or includes (EncryptionContextSubset) the encryption context specified in this structure. For more information about encryption context, see Encryption Context in the AWS Key Management Service Developer Guide . For information about grant constraints, see Using grant constraints in the AWS Key Management Service Developer Guide.

The encryption context grant constraints are supported only on operations that include an encryption context. You cannot use an encryption context grant constraint for cryptographic operations with asymmetric CMKs or for management operations, such as DescribeKey or RetireGrant.

pub fn set_constraints(self, input: Option<GrantConstraints>) -> Self

pub fn grant_tokens(self, input: impl Into<String>) -> Self

pub fn set_grant_tokens(self, input: Option<Vec<String>>) -> Self

pub fn name(self, input: impl Into<String>) -> Self

A friendly name for the grant. Use this value to prevent the unintended creation of duplicate grants when retrying this request.

When this value is absent, all CreateGrant requests result in a new grant with a unique GrantId even if all the supplied parameters are identical. This can result in unintended duplicates when you retry the CreateGrant request.

When this value is present, you can retry a CreateGrant request with identical parameters; if the grant already exists, the original GrantId is returned without creating a new grant. Note that the returned grant token is unique with every CreateGrant request, even when a duplicate GrantId is returned. All grant tokens for the same grant ID can be used interchangeably.

pub fn set_name(self, input: Option<String>) -> Self

pub fn build(self) -> Result<CreateGrantInput, BuildError>

Consumes the builder and constructs a CreateGrantInput