Struct aws_sdk_kms::input::create_key_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateKeyInput

Implementations

impl Builder

pub fn policy(self, input: impl Into<String>) -> Self

The key policy to attach to the CMK.

If you provide a key policy, it must meet the following criteria:

If you don't set BypassPolicyLockoutSafetyCheck to true, the key policy must allow the principal that is making the CreateKey request to make a subsequent PutKeyPolicy request on the CMK. This reduces the risk that the CMK becomes unmanageable. For more information, refer to the scenario in the Default Key Policy section of the AWS Key Management Service Developer Guide .

Each statement in the key policy must contain one or more principals. The principals in the key policy must exist and be visible to AWS KMS. When you create a new AWS principal (for example, an IAM user or role), you might need to enforce a delay before including the new principal in a key policy because the new principal might not be immediately visible to AWS KMS. For more information, see Changes that I make are not always immediately visible in the AWS Identity and Access Management User Guide.

If you do not provide a key policy, AWS KMS attaches a default key policy to the CMK. For more information, see Default Key Policy in the AWS Key Management Service Developer Guide.

The key policy size quota is 32 kilobytes (32768 bytes).

For help writing and formatting a JSON policy document, see the IAM JSON Policy Reference in the IAM User Guide .

pub fn set_policy(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A description of the CMK.

Use a description that helps you decide whether the CMK is appropriate for a task. The default value is an empty string (no description).

pub fn set_description(self, input: Option<String>) -> Self

pub fn key_usage(self, input: KeyUsageType) -> Self

Determines the cryptographic operations for which you can use the CMK. The default value is ENCRYPT_DECRYPT. This parameter is required only for asymmetric CMKs. You can't change the KeyUsage value after the CMK is created.

Select only one valid value.

For symmetric CMKs, omit the parameter or specify ENCRYPT_DECRYPT.

For asymmetric CMKs with RSA key material, specify ENCRYPT_DECRYPT or SIGN_VERIFY.

For asymmetric CMKs with ECC key material, specify SIGN_VERIFY.

pub fn set_key_usage(self, input: Option<KeyUsageType>) -> Self

pub fn customer_master_key_spec(self, input: CustomerMasterKeySpec) -> Self

Specifies the type of CMK to create. The default value, SYMMETRIC_DEFAULT, creates a CMK with a 256-bit symmetric key for encryption and decryption. For help choosing a key spec for your CMK, see How to Choose Your CMK Configuration in the AWS Key Management Service Developer Guide.

The CustomerMasterKeySpec determines whether the CMK contains a symmetric key or an asymmetric key pair. It also determines the encryption algorithms or signing algorithms that the CMK supports. You can't change the CustomerMasterKeySpec after the CMK is created. To further restrict the algorithms that can be used with the CMK, use a condition key in its key policy or IAM policy. For more information, see kms:EncryptionAlgorithm or kms:Signing Algorithm in the AWS Key Management Service Developer Guide.

AWS services that are integrated with AWS KMS use symmetric CMKs to protect your data. These services do not support asymmetric CMKs. For help determining whether a CMK is symmetric or asymmetric, see Identifying Symmetric and Asymmetric CMKs in the AWS Key Management Service Developer Guide.

AWS KMS supports the following key specs for CMKs:

Symmetric key (default)

SYMMETRIC_DEFAULT (AES-256-GCM)

Asymmetric RSA key pairs

RSA_2048

RSA_3072

RSA_4096

Asymmetric NIST-recommended elliptic curve key pairs

ECC_NIST_P256 (secp256r1)

ECC_NIST_P384 (secp384r1)

ECC_NIST_P521 (secp521r1)

Other asymmetric elliptic curve key pairs

ECC_SECG_P256K1 (secp256k1), commonly used for cryptocurrencies.

pub fn set_customer_master_key_spec(
self,
input: Option<CustomerMasterKeySpec>
) -> Self

pub fn origin(self, input: OriginType) -> Self

The source of the key material for the CMK. You cannot change the origin after you create the CMK. The default is AWS_KMS, which means that AWS KMS creates the key material.

To create a CMK with no key material (for imported key material), set the value to EXTERNAL. For more information about importing key material into AWS KMS, see Importing Key Material in the AWS Key Management Service Developer Guide. This value is valid only for symmetric CMKs.

To create a CMK in an AWS KMS custom key store and create its key material in the associated AWS CloudHSM cluster, set this value to AWS_CLOUDHSM. You must also use the CustomKeyStoreId parameter to identify the custom key store. This value is valid only for symmetric CMKs.

pub fn set_origin(self, input: Option<OriginType>) -> Self

pub fn custom_key_store_id(self, input: impl Into<String>) -> Self

Creates the CMK in the specified custom key store and the key material in its associated AWS CloudHSM cluster. To create a CMK in a custom key store, you must also specify the Origin parameter with a value of AWS_CLOUDHSM. The AWS CloudHSM cluster that is associated with the custom key store must have at least two active HSMs, each in a different Availability Zone in the Region.

This parameter is valid only for symmetric CMKs and regional CMKs. You cannot create an asymmetric CMK or a multi-Region CMK in a custom key store.

To find the ID of a custom key store, use the DescribeCustomKeyStores operation.

The response includes the custom key store ID and the ID of the AWS CloudHSM cluster.

This operation is part of the Custom Key Store feature feature in AWS KMS, which combines the convenience and extensive integration of AWS KMS with the isolation and control of a single-tenant key store.

pub fn set_custom_key_store_id(self, input: Option<String>) -> Self

pub fn bypass_policy_lockout_safety_check(self, input: bool) -> Self

A flag to indicate whether to bypass the key policy lockout safety check.

Setting this value to true increases the risk that the CMK becomes unmanageable. Do not set this value to true indiscriminately.

For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide .

Use this parameter only when you include a policy in the request and you intend to prevent the principal that is making the request from making a subsequent PutKeyPolicy request on the CMK.

The default value is false.

pub fn set_bypass_policy_lockout_safety_check(self, input: Option<bool>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn multi_region(self, input: bool) -> Self

Creates a multi-Region primary key that you can replicate into other AWS Regions. You cannot change this value after you create the CMK.

For a multi-Region key, set this parameter to True. For a single-Region CMK, omit this parameter or set it to False. The default value is False.

This operation supports multi-Region keys, an AWS KMS feature that lets you create multiple interoperable CMKs in different AWS Regions. Because these CMKs have the same key ID, key material, and other metadata, you can use them to encrypt data in one AWS Region and decrypt it in a different AWS Region without making a cross-Region call or exposing the plaintext data. For more information about multi-Region keys, see Using multi-Region keys in the AWS Key Management Service Developer Guide.

This value creates a primary key, not a replica. To create a replica key, use the ReplicateKey operation.

You can create a symmetric or asymmetric multi-Region CMK, and you can create a multi-Region CMK with imported key material. However, you cannot create a multi-Region CMK in a custom key store.

pub fn set_multi_region(self, input: Option<bool>) -> Self

pub fn build(self) -> Result<CreateKeyInput, BuildError>

Consumes the builder and constructs a CreateKeyInput