Struct aws_sdk_kms::input::CreateGrantInput

#[non_exhaustive]
pub struct CreateGrantInput {
    pub key_id: Option<String>,
    pub grantee_principal: Option<String>,
    pub retiring_principal: Option<String>,
    pub operations: Option<Vec<GrantOperation>>,
    pub constraints: Option<GrantConstraints>,
    pub grant_tokens: Option<Vec<String>>,
    pub name: Option<String>,
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

key_id: Option<String>

Identifies the customer master key (CMK) for the grant. The grant gives principals permission to use this CMK.

Specify the key ID or key ARN of the CMK. To specify a CMK in a different AWS account, you must use the key ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

To get the key ID and key ARN for a CMK, use ListKeys or DescribeKey.

grantee_principal: Option<String>

The identity that gets the permissions specified in the grant.

To specify the principal, use the Amazon Resource Name (ARN) of an AWS principal. Valid AWS principals include AWS accounts (root), IAM users, IAM roles, federated users, and assumed role users. For examples of the ARN syntax to use for specifying a principal, see AWS Identity and Access Management (IAM) in the Example ARNs section of the AWS General Reference.

retiring_principal: Option<String>

The principal that is given permission to retire the grant by using RetireGrant operation.

To specify the principal, use the Amazon Resource Name (ARN) of an AWS principal. Valid AWS principals include AWS accounts (root), IAM users, federated users, and assumed role users. For examples of the ARN syntax to use for specifying a principal, see AWS Identity and Access Management (IAM) in the Example ARNs section of the AWS General Reference.

operations: Option<Vec<GrantOperation>>

A list of operations that the grant permits.

The operation must be supported on the CMK. For example, you cannot create a grant for a symmetric CMK that allows the Sign operation, or a grant for an asymmetric CMK that allows the GenerateDataKey operation. If you try, AWS KMS returns a ValidationError exception. For details, see Grant operations in the AWS Key Management Service Developer Guide.

constraints: Option<GrantConstraints>

Specifies a grant constraint.

AWS KMS supports the EncryptionContextEquals and EncryptionContextSubset grant constraints. Each constraint value can include up to 8 encryption context pairs. The encryption context value in each constraint cannot exceed 384 characters.

These grant constraints allow a cryptographic operation only when the encryption context in the request matches (EncryptionContextEquals) or includes (EncryptionContextSubset) the encryption context specified in this structure. For more information about encryption context, see Encryption Context in the AWS Key Management Service Developer Guide . For information about grant constraints, see Using grant constraints in the AWS Key Management Service Developer Guide.

The encryption context grant constraints are supported only on operations that include an encryption context. You cannot use an encryption context grant constraint for cryptographic operations with asymmetric CMKs or for management operations, such as DescribeKey or RetireGrant.

grant_tokens: Option<Vec<String>>

A list of grant tokens.

Use a grant token when your permission to call this operation comes from a new grant that has not yet achieved eventual consistency. For more information, see Grant token in the AWS Key Management Service Developer Guide.

name: Option<String>

A friendly name for the grant. Use this value to prevent the unintended creation of duplicate grants when retrying this request.

When this value is absent, all CreateGrant requests result in a new grant with a unique GrantId even if all the supplied parameters are identical. This can result in unintended duplicates when you retry the CreateGrant request.

When this value is present, you can retry a CreateGrant request with identical parameters; if the grant already exists, the original GrantId is returned without creating a new grant. Note that the returned grant token is unique with every CreateGrant request, even when a duplicate GrantId is returned. All grant tokens for the same grant ID can be used interchangeably.

Implementations

impl CreateGrantInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<CreateGrant, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<CreateGrant>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture CreateGrantInput