Module aws_sdk_kms::model

Modules

alias_list_entry

See AliasListEntry

custom_key_stores_list_entry

See CustomKeyStoresListEntry

grant_constraints

See GrantConstraints

grant_list_entry

See GrantListEntry

key_list_entry

See KeyListEntry

key_metadata

See KeyMetadata

multi_region_configuration

See MultiRegionConfiguration

multi_region_key

See MultiRegionKey

tag

See Tag

Structs

AliasListEntry

Contains information about an alias.

CustomKeyStoresListEntry

Contains information about each custom key store in the custom key store list.

GrantConstraints

Use this structure to allow cryptographic operations in the grant only when the operation request

GrantListEntry

Contains information about a grant.

KeyListEntry

Contains information about each entry in the key list.

KeyMetadata

Contains metadata about a customer master key (CMK).

MultiRegionConfiguration

Describes the configuration of this multi-Region CMK. This field appears only when the CMK

MultiRegionKey

Describes the primary or replica key in a multi-Region key.

Tag

A key-value pair. A tag consists of a tag key and a tag value. Tag keys and tag values are

Enums

AlgorithmSpec

ConnectionErrorCodeType

ConnectionStateType

CustomerMasterKeySpec

DataKeyPairSpec

DataKeySpec

EncryptionAlgorithmSpec

ExpirationModelType

GrantOperation

KeyManagerType

KeyState

KeyUsageType

MessageType

MultiRegionKeyType

OriginType

SigningAlgorithmSpec

WrappingKeySpec