Struct aws_sdk_kms::model::key_metadata::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for KeyMetadata

Implementations

impl Builder

pub fn aws_account_id(self, input: impl Into<String>) -> Self

The twelve-digit account ID of the Amazon Web Services account that owns the KMS key.

pub fn set_aws_account_id(self, input: Option<String>) -> Self

pub fn key_id(self, input: impl Into<String>) -> Self

The globally unique identifier for the KMS key.

pub fn set_key_id(self, input: Option<String>) -> Self

pub fn arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the KMS key. For examples, see Key Management Service (KMS) in the Example ARNs section of the Amazon Web Services General Reference.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn creation_date(self, input: Instant) -> Self

The date and time when the KMS key was created.

pub fn set_creation_date(self, input: Option<Instant>) -> Self

pub fn enabled(self, input: bool) -> Self

Specifies whether the KMS key is enabled. When KeyState is Enabled this value is true, otherwise it is false.

pub fn set_enabled(self, input: Option<bool>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

The description of the KMS key.

pub fn set_description(self, input: Option<String>) -> Self

pub fn key_usage(self, input: KeyUsageType) -> Self

The cryptographic operations for which you can use the KMS key.

pub fn set_key_usage(self, input: Option<KeyUsageType>) -> Self

pub fn key_state(self, input: KeyState) -> Self

The current status of the KMS key.

For more information about how key state affects the use of a KMS key, see Key state: Effect on your KMS key in the Key Management Service Developer Guide.

pub fn set_key_state(self, input: Option<KeyState>) -> Self

pub fn deletion_date(self, input: Instant) -> Self

The date and time after which KMS deletes this KMS key. This value is present only when the KMS key is scheduled for deletion, that is, when its KeyState is PendingDeletion.

When the primary key in a multi-Region key is scheduled for deletion but still has replica keys, its key state is PendingReplicaDeletion and the length of its waiting period is displayed in the PendingDeletionWindowInDays field.

pub fn set_deletion_date(self, input: Option<Instant>) -> Self

pub fn valid_to(self, input: Instant) -> Self

The time at which the imported key material expires. When the key material expires, KMS deletes the key material and the KMS key becomes unusable. This value is present only for KMS keys whose Origin is EXTERNAL and whose ExpirationModel is KEY_MATERIAL_EXPIRES, otherwise this value is omitted.

pub fn set_valid_to(self, input: Option<Instant>) -> Self

pub fn origin(self, input: OriginType) -> Self

The source of the key material for the KMS key. When this value is AWS_KMS, KMS created the key material. When this value is EXTERNAL, the key material was imported or the KMS key doesn't have any key material. When this value is AWS_CLOUDHSM, the key material was created in the CloudHSM cluster associated with a custom key store.

pub fn set_origin(self, input: Option<OriginType>) -> Self

pub fn custom_key_store_id(self, input: impl Into<String>) -> Self

A unique identifier for the custom key store that contains the KMS key. This value is present only when the KMS key is created in a custom key store.

pub fn set_custom_key_store_id(self, input: Option<String>) -> Self

pub fn cloud_hsm_cluster_id(self, input: impl Into<String>) -> Self

The cluster ID of the CloudHSM cluster that contains the key material for the KMS key. When you create a KMS key in a custom key store, KMS creates the key material for the KMS key in the associated CloudHSM cluster. This value is present only when the KMS key is created in a custom key store.

pub fn set_cloud_hsm_cluster_id(self, input: Option<String>) -> Self

pub fn expiration_model(self, input: ExpirationModelType) -> Self

Specifies whether the KMS key's key material expires. This value is present only when Origin is EXTERNAL, otherwise this value is omitted.

pub fn set_expiration_model(self, input: Option<ExpirationModelType>) -> Self

pub fn key_manager(self, input: KeyManagerType) -> Self

The manager of the KMS key. KMS keys in your Amazon Web Services account are either customer managed or Amazon Web Services managed. For more information about the difference, see KMS keys in the Key Management Service Developer Guide.

pub fn set_key_manager(self, input: Option<KeyManagerType>) -> Self

pub fn customer_master_key_spec(self, input: CustomerMasterKeySpec) -> Self

Instead, use the KeySpec field.

The KeySpec and CustomerMasterKeySpec fields have the same value. We recommend that you use the KeySpec field in your code. However, to avoid breaking changes, KMS will support both fields.

pub fn set_customer_master_key_spec(
self,
input: Option<CustomerMasterKeySpec>
) -> Self

pub fn key_spec(self, input: KeySpec) -> Self

Describes the type of key material in the KMS key.

pub fn set_key_spec(self, input: Option<KeySpec>) -> Self

pub fn encryption_algorithms(
self,
input: impl Into<EncryptionAlgorithmSpec>
) -> Self

pub fn set_encryption_algorithms(
self,
input: Option<Vec<EncryptionAlgorithmSpec>>
) -> Self

pub fn signing_algorithms(self, input: impl Into<SigningAlgorithmSpec>) -> Self

pub fn set_signing_algorithms(
self,
input: Option<Vec<SigningAlgorithmSpec>>
) -> Self

pub fn multi_region(self, input: bool) -> Self

Indicates whether the KMS key is a multi-Region (True) or regional (False) key. This value is True for multi-Region primary and replica keys and False for regional KMS keys.

For more information about multi-Region keys, see Using multi-Region keys in the Key Management Service Developer Guide.

pub fn set_multi_region(self, input: Option<bool>) -> Self

pub fn multi_region_configuration(self, input: MultiRegionConfiguration) -> Self

Lists the primary and replica keys in same multi-Region key. This field is present only when the value of the MultiRegion field is True.

For more information about any listed KMS key, use the DescribeKey operation.

MultiRegionKeyType indicates whether the KMS key is a PRIMARY or REPLICA key.

PrimaryKey displays the key ARN and Region of the primary key. This field displays the current KMS key if it is the primary key.

ReplicaKeys displays the key ARNs and Regions of all replica keys. This field includes the current KMS key if it is a replica key.

pub fn set_multi_region_configuration(
self,
input: Option<MultiRegionConfiguration>
) -> Self

pub fn pending_deletion_window_in_days(self, input: i32) -> Self

The waiting period before the primary key in a multi-Region key is deleted. This waiting period begins when the last of its replica keys is deleted. This value is present only when the KeyState of the KMS key is PendingReplicaDeletion. That indicates that the KMS key is the primary key in a multi-Region key, it is scheduled for deletion, and it still has existing replica keys.

When a single-Region KMS key or a multi-Region replica key is scheduled for deletion, its deletion date is displayed in the DeletionDate field. However, when the primary key in a multi-Region key is scheduled for deletion, its waiting period doesn't begin until all of its replica keys are deleted. This value displays that waiting period. When the last replica key in the multi-Region key is deleted, the KeyState of the scheduled primary key changes from PendingReplicaDeletion to PendingDeletion and the deletion date appears in the DeletionDate field.

pub fn set_pending_deletion_window_in_days(self, input: Option<i32>) -> Self

pub fn build(self) -> KeyMetadata

Consumes the builder and constructs a KeyMetadata