Struct aws_sdk_kms::model::KeyMetadata

#[non_exhaustive]
pub struct KeyMetadata {

    pub aws_account_id: Option<String>,
    pub key_id: Option<String>,
    pub arn: Option<String>,
    pub creation_date: Option<Instant>,
    pub enabled: bool,
    pub description: Option<String>,
    pub key_usage: Option<KeyUsageType>,
    pub key_state: Option<KeyState>,
    pub deletion_date: Option<Instant>,
    pub valid_to: Option<Instant>,
    pub origin: Option<OriginType>,
    pub custom_key_store_id: Option<String>,
    pub cloud_hsm_cluster_id: Option<String>,
    pub expiration_model: Option<ExpirationModelType>,
    pub key_manager: Option<KeyManagerType>,
    pub customer_master_key_spec: Option<CustomerMasterKeySpec>,
    pub encryption_algorithms: Option<Vec<EncryptionAlgorithmSpec>>,
    pub signing_algorithms: Option<Vec<SigningAlgorithmSpec>>,
    pub multi_region: Option<bool>,
    pub multi_region_configuration: Option<MultiRegionConfiguration>,
    pub pending_deletion_window_in_days: Option<i32>,

}

Contains metadata about a customer master key (CMK).

This data type is used as a response element for the CreateKey and DescribeKey operations.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

aws_account_id: Option<String>

The twelve-digit account ID of the AWS account that owns the CMK.

key_id: Option<String>

The globally unique identifier for the CMK.

arn: Option<String>

The Amazon Resource Name (ARN) of the CMK. For examples, see AWS Key Management Service (AWS KMS) in the Example ARNs section of the AWS General Reference.

creation_date: Option<Instant>

The date and time when the CMK was created.

enabled: bool

Specifies whether the CMK is enabled. When KeyState is Enabled this value is true, otherwise it is false.

description: Option<String>

The description of the CMK.

key_usage: Option<KeyUsageType>

The cryptographic operations for which you can use the CMK.

key_state: Option<KeyState>

The current status of the CMK.

For more information about how key state affects the use of a CMK, see Key state: Effect on your CMK in the AWS Key Management Service Developer Guide.

deletion_date: Option<Instant>

The date and time after which AWS KMS deletes this CMK. This value is present only when the CMK is scheduled for deletion, that is, when its KeyState is PendingDeletion.

When the primary key in a multi-Region key is scheduled for deletion but still has replica keys, its key state is PendingReplicaDeletion and the length of its waiting period is displayed in the PendingDeletionWindowInDays field.

valid_to: Option<Instant>

The time at which the imported key material expires. When the key material expires, AWS KMS deletes the key material and the CMK becomes unusable. This value is present only for CMKs whose Origin is EXTERNAL and whose ExpirationModel is KEY_MATERIAL_EXPIRES, otherwise this value is omitted.

origin: Option<OriginType>

The source of the CMK's key material. When this value is AWS_KMS, AWS KMS created the key material. When this value is EXTERNAL, the key material was imported from your existing key management infrastructure or the CMK lacks key material. When this value is AWS_CLOUDHSM, the key material was created in the AWS CloudHSM cluster associated with a custom key store.

custom_key_store_id: Option<String>

A unique identifier for the custom key store that contains the CMK. This value is present only when the CMK is created in a custom key store.

cloud_hsm_cluster_id: Option<String>

The cluster ID of the AWS CloudHSM cluster that contains the key material for the CMK. When you create a CMK in a custom key store, AWS KMS creates the key material for the CMK in the associated AWS CloudHSM cluster. This value is present only when the CMK is created in a custom key store.

expiration_model: Option<ExpirationModelType>

Specifies whether the CMK's key material expires. This value is present only when Origin is EXTERNAL, otherwise this value is omitted.

key_manager: Option<KeyManagerType>

The manager of the CMK. CMKs in your AWS account are either customer managed or AWS managed. For more information about the difference, see Customer Master Keys in the AWS Key Management Service Developer Guide.

customer_master_key_spec: Option<CustomerMasterKeySpec>

Describes the type of key material in the CMK.

encryption_algorithms: Option<Vec<EncryptionAlgorithmSpec>>

The encryption algorithms that the CMK supports. You cannot use the CMK with other encryption algorithms within AWS KMS.

This value is present only when the KeyUsage of the CMK is ENCRYPT_DECRYPT.

signing_algorithms: Option<Vec<SigningAlgorithmSpec>>

The signing algorithms that the CMK supports. You cannot use the CMK with other signing algorithms within AWS KMS.

This field appears only when the KeyUsage of the CMK is SIGN_VERIFY.

multi_region: Option<bool>

Indicates whether the CMK is a multi-Region (True) or regional (False) key. This value is True for multi-Region primary and replica CMKs and False for regional CMKs.

For more information about multi-Region keys, see Using multi-Region keys in the AWS Key Management Service Developer Guide.

multi_region_configuration: Option<MultiRegionConfiguration>

Lists the primary and replica CMKs in same multi-Region CMK. This field is present only when the value of the MultiRegion field is True.

For more information about any listed CMK, use the DescribeKey operation.

MultiRegionKeyType indicates whether the CMK is a PRIMARY or REPLICA key.

PrimaryKey displays the key ARN and Region of the primary key. This field displays the current CMK if it is the primary key.

ReplicaKeys displays the key ARNs and Regions of all replica keys. This field includes the current CMK if it is a replica key.

pending_deletion_window_in_days: Option<i32>

The waiting period before the primary key in a multi-Region key is deleted. This waiting period begins when the last of its replica keys is deleted. This value is present only when the KeyState of the CMK is PendingReplicaDeletion. That indicates that the CMK is the primary key in a multi-Region key, it is scheduled for deletion, and it still has existing replica keys.

When a regional CMK or a replica key in a multi-Region key is scheduled for deletion, its deletion date is displayed in the DeletionDate field. However, when the primary key in a multi-Region key is scheduled for deletion, its waiting period doesn't begin until all of its replica keys are deleted. This value displays that waiting period. When the last replica key in the multi-Region key is deleted, the KeyState of the scheduled primary key changes from PendingReplicaDeletion to PendingDeletion and the deletion date appears in the DeletionDate field.

Implementations

impl KeyMetadata

pub fn builder() -> Builder

Creates a new builder-style object to manufacture KeyMetadata