Module aws_sdk_kms::operation

Structs

CancelKeyDeletion

Cancels the deletion of a customer master key (CMK). When this operation succeeds, the key

ConnectCustomKeyStore

Connects or reconnects a custom key store to its associated AWS CloudHSM cluster.

CreateAlias

Creates a friendly name for a customer master key (CMK).

CreateCustomKeyStore

Creates a custom key store that is associated with an AWS CloudHSM cluster that you own and

CreateGrant

Adds a grant to a customer master key (CMK).

CreateKey

Creates a unique customer managed customer master key (CMK) in your AWS

Decrypt

Decrypts ciphertext that was encrypted by a AWS KMS customer master key (CMK) using any of

DeleteAlias

Deletes the specified alias.

DeleteCustomKeyStore

Deletes a custom key store. This operation does not delete the AWS CloudHSM cluster that is

DeleteImportedKeyMaterial

Deletes key material that you previously imported. This operation makes the specified

DescribeCustomKeyStores

Gets information about custom key stores in the account and Region.

DescribeKey

Provides detailed information about a customer master key (CMK). You can run

DisableKey

Sets the state of a customer master key (CMK) to disabled. This change temporarily

DisableKeyRotation

Disables automatic

DisconnectCustomKeyStore

Disconnects the custom key store from its associated AWS CloudHSM cluster. While a custom key

EnableKey

Sets the key state of a customer master key (CMK) to enabled. This allows you to use the

EnableKeyRotation

Enables automatic rotation

Encrypt

Encrypts plaintext into ciphertext by using a customer master key (CMK). The

GenerateDataKey

Generates a unique symmetric data key for client-side encryption. This operation returns a

GenerateDataKeyPair

Generates a unique asymmetric data key pair. The GenerateDataKeyPair

GenerateDataKeyPairWithoutPlaintext

Generates a unique asymmetric data key pair. The

GenerateDataKeyWithoutPlaintext

Generates a unique symmetric data key. This operation returns a data key that is encrypted

GenerateRandom

Returns a random byte string that is cryptographically secure.

GetKeyPolicy

Gets a key policy attached to the specified customer master key (CMK).

GetKeyRotationStatus

Gets a Boolean value that indicates whether automatic rotation of the key material is

GetParametersForImport

Returns the items you need to import key material into a symmetric, customer managed

GetPublicKey

Returns the public key of an asymmetric CMK. Unlike the private key of a asymmetric CMK,

ImportKeyMaterial

Imports key material into an existing symmetric AWS KMS customer master key (CMK) that was

ListAliases

Gets a list of aliases in the caller's AWS account and region. For more information about

ListGrants

Gets a list of all grants for the specified customer master key (CMK).

ListKeyPolicies

Gets the names of the key policies that are attached to a customer master key (CMK). This

ListKeys

Gets a list of all customer master keys (CMKs) in the caller's AWS account and

ListResourceTags

Returns all tags on the specified customer master key (CMK).

ListRetirableGrants

Returns information about all grants in the AWS account and Region that have the specified

PutKeyPolicy

Attaches a key policy to the specified customer master key (CMK).

ReEncrypt

Decrypts ciphertext and then reencrypts it entirely within AWS KMS. You can use this

ReplicateKey

Replicates a multi-Region key into the specified Region. This operation creates a

RetireGrant

Deletes a grant. Typically, you retire a grant when you no longer need its permissions. To

RevokeGrant

Deletes the specified grant. You revoke a grant to terminate the permissions that the

ScheduleKeyDeletion

Schedules the deletion of a customer master key (CMK). By default, AWS KMS applies a waiting

Sign

Creates a digital

TagResource

Adds or edits tags on a customer managed CMK.

UntagResource

Deletes tags from a customer managed CMK. To delete a tag,

UpdateAlias

Associates an existing AWS KMS alias with a different customer master key (CMK). Each alias

UpdateCustomKeyStore

Changes the properties of a custom key store. Use the CustomKeyStoreId

UpdateKeyDescription

Updates the description of a customer master key (CMK). To see the description of a CMK,

UpdatePrimaryRegion

Changes the primary key of a multi-Region key.

Verify

Verifies a digital signature that was generated by the Sign operation.