Module aws_sdk_kms::error

Modules

already_exists_error

See AlreadyExistsError

cloud_hsm_cluster_in_use_error

See CloudHsmClusterInUseError

cloud_hsm_cluster_invalid_configuration_error

See CloudHsmClusterInvalidConfigurationError

cloud_hsm_cluster_not_active_error

See CloudHsmClusterNotActiveError

cloud_hsm_cluster_not_found_error

See CloudHsmClusterNotFoundError

cloud_hsm_cluster_not_related_error

See CloudHsmClusterNotRelatedError

custom_key_store_has_cm_ks_error

See CustomKeyStoreHasCMKsError

custom_key_store_invalid_state_error

See CustomKeyStoreInvalidStateError

custom_key_store_name_in_use_error

See CustomKeyStoreNameInUseError

custom_key_store_not_found_error

See CustomKeyStoreNotFoundError

dependency_timeout_error

See DependencyTimeoutError

disabled_error

See DisabledError

expired_import_token_error

See ExpiredImportTokenError

incorrect_key_error

See IncorrectKeyError

incorrect_key_material_error

See IncorrectKeyMaterialError

incorrect_trust_anchor_error

See IncorrectTrustAnchorError

invalid_alias_name_error

See InvalidAliasNameError

invalid_arn_error

See InvalidArnError

invalid_ciphertext_error

See InvalidCiphertextError

invalid_grant_id_error

See InvalidGrantIdError

invalid_grant_token_error

See InvalidGrantTokenError

invalid_import_token_error

See InvalidImportTokenError

invalid_key_usage_error

See InvalidKeyUsageError

invalid_marker_error

See InvalidMarkerError

key_unavailable_error

See KeyUnavailableError

kms_internal_error

See KMSInternalError

kms_invalid_signature_error

See KMSInvalidSignatureError

kms_invalid_state_error

See KMSInvalidStateError

limit_exceeded_error

See LimitExceededError

malformed_policy_document_error

See MalformedPolicyDocumentError

not_found_error

See NotFoundError

tag_error

See TagError

unsupported_operation_error

See UnsupportedOperationError

Structs

AlreadyExistsError

The request was rejected because it attempted to create a resource that already

CancelKeyDeletionError

CloudHsmClusterInUseError

The request was rejected because the specified AWS CloudHSM cluster is already associated with a

CloudHsmClusterInvalidConfigurationError

The request was rejected because the associated AWS CloudHSM cluster did not meet the

CloudHsmClusterNotActiveError

The request was rejected because the AWS CloudHSM cluster that is associated with the custom key

CloudHsmClusterNotFoundError

The request was rejected because AWS KMS cannot find the AWS CloudHSM cluster with the specified

CloudHsmClusterNotRelatedError

The request was rejected because the specified AWS CloudHSM cluster has a different cluster

ConnectCustomKeyStoreError

CreateAliasError

CreateCustomKeyStoreError

CreateGrantError

CreateKeyError

CustomKeyStoreHasCMKsError

The request was rejected because the custom key store contains AWS KMS customer master keys

CustomKeyStoreInvalidStateError

The request was rejected because of the ConnectionState of the custom key

CustomKeyStoreNameInUseError

The request was rejected because the specified custom key store name is already assigned

CustomKeyStoreNotFoundError

The request was rejected because AWS KMS cannot find a custom key store with the specified

DecryptError

DeleteAliasError

DeleteCustomKeyStoreError

DeleteImportedKeyMaterialError

DependencyTimeoutError

The system timed out while trying to fulfill the request. The request can be

DescribeCustomKeyStoresError

DescribeKeyError

DisableKeyError

DisableKeyRotationError

DisabledError

The request was rejected because the specified CMK is not enabled.

DisconnectCustomKeyStoreError

EnableKeyError

EnableKeyRotationError

EncryptError

ExpiredImportTokenError

The request was rejected because the specified import token is expired. Use GetParametersForImport to get a new import token and public key, use the new

GenerateDataKeyError

GenerateDataKeyPairError

GenerateDataKeyPairWithoutPlaintextError

GenerateDataKeyWithoutPlaintextError

GenerateRandomError

GetKeyPolicyError

GetKeyRotationStatusError

GetParametersForImportError

GetPublicKeyError

ImportKeyMaterialError

IncorrectKeyError

The request was rejected because the specified CMK cannot decrypt the data. The

IncorrectKeyMaterialError

The request was rejected because the key material in the request is, expired, invalid, or

IncorrectTrustAnchorError

The request was rejected because the trust anchor certificate in the request is not the

InvalidAliasNameError

The request was rejected because the specified alias name is not valid.

InvalidArnError

The request was rejected because a specified ARN, or an ARN in a key policy, is not

InvalidCiphertextError

From the Decrypt or ReEncrypt operation, the request

InvalidGrantIdError

The request was rejected because the specified GrantId is not valid.

InvalidGrantTokenError

The request was rejected because the specified grant token is not valid.

InvalidImportTokenError

The request was rejected because the provided import token is invalid or is associated

InvalidKeyUsageError

The request was rejected for one of the following reasons:

InvalidMarkerError

The request was rejected because the marker that specifies where pagination should next

KMSInternalError

The request was rejected because an internal exception occurred. The request can be

KMSInvalidSignatureError

The request was rejected because the signature verification failed. Signature

KMSInvalidStateError

The request was rejected because the state of the specified resource is not valid for this

KeyUnavailableError

The request was rejected because the specified CMK was not available. You can retry the

LimitExceededError

The request was rejected because a quota was exceeded. For more information, see Quotas in the

ListAliasesError

ListGrantsError

ListKeyPoliciesError

ListKeysError

ListResourceTagsError

ListRetirableGrantsError

MalformedPolicyDocumentError

The request was rejected because the specified policy is not syntactically or semantically

NotFoundError

The request was rejected because the specified entity or resource could not be

PutKeyPolicyError

ReEncryptError

RetireGrantError

RevokeGrantError

ScheduleKeyDeletionError

SignError

TagError

The request was rejected because one or more tags are not valid.

TagResourceError

UnsupportedOperationError

The request was rejected because a specified parameter is not supported or a specified

UntagResourceError

UpdateAliasError

UpdateCustomKeyStoreError

UpdateKeyDescriptionError

VerifyError

Enums

CancelKeyDeletionErrorKind

ConnectCustomKeyStoreErrorKind

CreateAliasErrorKind

CreateCustomKeyStoreErrorKind

CreateGrantErrorKind

CreateKeyErrorKind

DecryptErrorKind

DeleteAliasErrorKind

DeleteCustomKeyStoreErrorKind

DeleteImportedKeyMaterialErrorKind

DescribeCustomKeyStoresErrorKind

DescribeKeyErrorKind

DisableKeyErrorKind

DisableKeyRotationErrorKind

DisconnectCustomKeyStoreErrorKind

EnableKeyErrorKind

EnableKeyRotationErrorKind

EncryptErrorKind

GenerateDataKeyErrorKind

GenerateDataKeyPairErrorKind

GenerateDataKeyPairWithoutPlaintextErrorKind

GenerateDataKeyWithoutPlaintextErrorKind

GenerateRandomErrorKind

GetKeyPolicyErrorKind

GetKeyRotationStatusErrorKind

GetParametersForImportErrorKind

GetPublicKeyErrorKind

ImportKeyMaterialErrorKind

ListAliasesErrorKind

ListGrantsErrorKind

ListKeyPoliciesErrorKind

ListKeysErrorKind

ListResourceTagsErrorKind

ListRetirableGrantsErrorKind

PutKeyPolicyErrorKind

ReEncryptErrorKind

RetireGrantErrorKind

RevokeGrantErrorKind

ScheduleKeyDeletionErrorKind

SignErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAliasErrorKind

UpdateCustomKeyStoreErrorKind

UpdateKeyDescriptionErrorKind

VerifyErrorKind