Struct aws_sdk_kms::model::key_metadata::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for KeyMetadata

Implementations

impl Builder

pub fn aws_account_id(self, inp: impl Into<String>) -> Self

The twelve-digit account ID of the AWS account that owns the CMK.

pub fn set_aws_account_id(self, inp: Option<String>) -> Self

pub fn key_id(self, inp: impl Into<String>) -> Self

The globally unique identifier for the CMK.

pub fn set_key_id(self, inp: Option<String>) -> Self

pub fn arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the CMK. For examples, see AWS Key Management Service (AWS KMS) in the Example ARNs section of the AWS General Reference.

pub fn set_arn(self, inp: Option<String>) -> Self

pub fn creation_date(self, inp: Instant) -> Self

The date and time when the CMK was created.

pub fn set_creation_date(self, inp: Option<Instant>) -> Self

pub fn enabled(self, inp: bool) -> Self

Specifies whether the CMK is enabled. When KeyState is Enabled this value is true, otherwise it is false.

pub fn set_enabled(self, inp: bool) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

The description of the CMK.

pub fn set_description(self, inp: Option<String>) -> Self

pub fn key_usage(self, inp: KeyUsageType) -> Self

The cryptographic operations for which you can use the CMK.

pub fn set_key_usage(self, inp: Option<KeyUsageType>) -> Self

pub fn key_state(self, inp: KeyState) -> Self

The current status of the CMK.

For more information about how key state affects the use of a CMK, see Key state: Effect on your CMK in the AWS Key Management Service Developer Guide.

pub fn set_key_state(self, inp: Option<KeyState>) -> Self

pub fn deletion_date(self, inp: Instant) -> Self

The date and time after which AWS KMS deletes the CMK. This value is present only when KeyState is PendingDeletion.

pub fn set_deletion_date(self, inp: Option<Instant>) -> Self

pub fn valid_to(self, inp: Instant) -> Self

The time at which the imported key material expires. When the key material expires, AWS KMS deletes the key material and the CMK becomes unusable. This value is present only for CMKs whose Origin is EXTERNAL and whose ExpirationModel is KEY_MATERIAL_EXPIRES, otherwise this value is omitted.

pub fn set_valid_to(self, inp: Option<Instant>) -> Self

pub fn origin(self, inp: OriginType) -> Self

The source of the CMK's key material. When this value is AWS_KMS, AWS KMS created the key material. When this value is EXTERNAL, the key material was imported from your existing key management infrastructure or the CMK lacks key material. When this value is AWS_CLOUDHSM, the key material was created in the AWS CloudHSM cluster associated with a custom key store.

pub fn set_origin(self, inp: Option<OriginType>) -> Self

pub fn custom_key_store_id(self, inp: impl Into<String>) -> Self

A unique identifier for the custom key store that contains the CMK. This value is present only when the CMK is created in a custom key store.

pub fn set_custom_key_store_id(self, inp: Option<String>) -> Self

pub fn cloud_hsm_cluster_id(self, inp: impl Into<String>) -> Self

The cluster ID of the AWS CloudHSM cluster that contains the key material for the CMK. When you create a CMK in a custom key store, AWS KMS creates the key material for the CMK in the associated AWS CloudHSM cluster. This value is present only when the CMK is created in a custom key store.

pub fn set_cloud_hsm_cluster_id(self, inp: Option<String>) -> Self

pub fn expiration_model(self, inp: ExpirationModelType) -> Self

Specifies whether the CMK's key material expires. This value is present only when Origin is EXTERNAL, otherwise this value is omitted.

pub fn set_expiration_model(self, inp: Option<ExpirationModelType>) -> Self

pub fn key_manager(self, inp: KeyManagerType) -> Self

The manager of the CMK. CMKs in your AWS account are either customer managed or AWS managed. For more information about the difference, see Customer Master Keys in the AWS Key Management Service Developer Guide.

pub fn set_key_manager(self, inp: Option<KeyManagerType>) -> Self

pub fn customer_master_key_spec(self, inp: CustomerMasterKeySpec) -> Self

Describes the type of key material in the CMK.

pub fn set_customer_master_key_spec(
self,
inp: Option<CustomerMasterKeySpec>
) -> Self

pub fn encryption_algorithms(
self,
inp: impl Into<EncryptionAlgorithmSpec>
) -> Self

pub fn set_encryption_algorithms(
self,
inp: Option<Vec<EncryptionAlgorithmSpec>>
) -> Self

pub fn signing_algorithms(self, inp: impl Into<SigningAlgorithmSpec>) -> Self

pub fn set_signing_algorithms(
self,
inp: Option<Vec<SigningAlgorithmSpec>>
) -> Self

pub fn build(self) -> KeyMetadata

Consumes the builder and constructs a KeyMetadata